Set integer type to unsigned and change item index type to int32 from int8

+++ Source/ElementusInventory/Public/ElementusInventoryComponent.h
@@ -35,7 +35,7 @@ struct FItemModifierData
 	}
 
 	FElementusItemInfo ItemInfo = FElementusItemInfo();	
-	int8 Index = INDEX_NONE;
+	int32 Index = INDEX_NONE;
 };
 
 DECLARE_DYNAMIC_MULTICAST_DELEGATE(FElementusInventoryUpdate);
